+++ src/app.module.ts
@@ -30,8 +30,8 @@ import { CrawlingModule } from './crawling/crawling.module';
         username: process.env.DB_USERNAME,
         password: process.env.DB_PASSWORD,
         database: process.env.DB_DATABASE,
-        entities: [__dirname + '/**/*.entity{.ts,.js}'],
-        // autoLoadEntities: true,
+        // entities: [__dirname + '/**/*.entity{.ts,.js}'],
+        autoLoadEntities: true,
         synchronize: true,
         logging: false,
         namingStrategy: new SnakeNamingStrategy(),
