+++ src/collection/collection.module.ts
@@ -5,6 +5,9 @@ import { CollectionRepository } from 'src/collection/repositories/collection.rep
 import { CollectionPlaceRepository } from 'src/collection/repositories/collection-place.repository';
 import { PlaceModule } from 'src/place/place.module';
 import { UserModule } from 'src/user/user.module';
+import { TypeOrmModule } from '@nestjs/typeorm';
+import { Collection } from './entities/collection.entity';
+import { CollectionPlace } from './entities/collection-place.entity';
 
 @Module({
   controllers: [CollectionController],
@@ -13,7 +16,11 @@ import { UserModule } from 'src/user/user.module';
     CollectionRepository,
     CollectionPlaceRepository,
   ],
-  imports: [PlaceModule, UserModule],
+  imports: [
+    TypeOrmModule.forFeature([Collection, CollectionPlace]),
+    PlaceModule,
+    UserModule,
+  ],
   exports: [CollectionService],
 })
 export class CollectionModule {}
