+++ src/place/place.module.ts
@@ -2,11 +2,18 @@ import { Module } from '@nestjs/common';
 import { PlaceController } from './place.controller';
 import { PlaceService } from './place.service';
 import { PlaceRepository } from 'src/place/repositories/place.repository';
-import { PlaceTagRepository } from 'src/repositories/place-tag.repository';
+
 import { PlaceImageRepository } from 'src/place/repositories/place-image.repository';
-import { PlaceDetailRepository } from 'src/repositories/place-detail.repository';
+
 import { TagModule } from 'src/tag/tag.module';
 import { S3Service } from 'src/place/s3.service';
+import { PlaceTagRepository } from './repositories/place-tag.repository';
+import { PlaceDetailRepository } from './repositories/place-detail.repository';
+import { TypeOrmModule } from '@nestjs/typeorm';
+import { Place } from './entities/place.entity';
+import { PlaceDetail } from './entities/place-detail.entity';
+import { PlaceImage } from './entities/place-image.entity';
+import { PlaceTag } from './entities/place-tag.entity';
 
 @Module({
   controllers: [PlaceController],
@@ -18,7 +25,10 @@ import { S3Service } from 'src/place/s3.service';
     PlaceDetailRepository,
     S3Service,
   ],
-  imports: [TagModule],
+  imports: [
+    TagModule,
+    TypeOrmModule.forFeature([Place, PlaceDetail, PlaceImage, PlaceTag]),
+  ],
   exports: [PlaceService],
 })
 export class PlaceModule {}
