+++ src/trip/trip.module.ts
@@ -2,11 +2,11 @@ import { Module } from '@nestjs/common';
 import { TripController } from './trip.controller';
 import { TripService } from './trip.service';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { Trip } from 'src/entities/trip.entity';
-import { TripStyle } from 'src/entities/trip-style.entity';
-import { AccommodationSchedule } from 'src/entities/accomodation-schedule.entity';
-import { PlaceSchedule } from 'src/entities/place-schedule.entity';
 import { TripRepository } from 'src/trip/repositories/trip.repository';
+import { Trip } from './entities/trip.entity';
+import { TripStyle } from './entities/trip-style.entity';
+import { AccommodationSchedule } from './entities/accomodation-schedule.entity';
+import { PlaceSchedule } from './entities/place-schedule.entity';
 
 @Module({
   imports: [
@@ -19,5 +19,6 @@ import { TripRepository } from 'src/trip/repositories/trip.repository';
   ],
   controllers: [TripController],
   providers: [TripService, TripRepository],
+  exports: [TripService],
 })
 export class TripModule {}
