+++ src/user/entities/preference.entity.ts
@@ -6,7 +6,7 @@ import {
   PrimaryGeneratedColumn,
   RelationId,
 } from 'typeorm';
-import { User } from '../user/entities/user.entity';
+import { User } from './user.entity';
 
 @Entity()
 export class Preference {
