+++ src/user/entities/user.entity.ts
@@ -9,11 +9,11 @@ import {
   RelationId,
 } from 'typeorm';
 import { Trip } from '../../trip/entities/trip.entity';
-import { Preference } from '../../common/entities/preference.entity';
 import { Folder } from '../../folder/entities/folder.entity';
 import { OAuthPlatform } from 'src/common/enums/oAuth-platform.enum';
 import { BaseEntity } from '../../common/entities/base-entity.entity';
-import { Collection } from './collection.entity';
+import { Preference } from './preference.entity';
+import { Collection } from 'src/collection/entities/collection.entity';
 
 @Entity()
 export class User extends BaseEntity {
